requirements.txt
langchain
langgraph
langchain-core
langchain-community
python-dotenv
langchain-openai
ipykernel
langchain-huggingface
sentence-transformers
langchain-groq
langgraph-cli[inmem]
faiss-cpu
tavily-python
streamlit
langchain-tavily
pypdf
mem0ai
torchvision
google-generativeai

